+++ cli.go
@@ -24,10 +24,25 @@ func (c *cli) Run(args []string) error {
 		completionScriptBash  = flag.Bool("completion-script-bash", false, "print completion-script-bash")
 		completionScriptZsh   = flag.Bool("completion-script-zsh", false, "print completion-script-zsh")
 		completionScriptClink = flag.Bool("completion-script-clink", false, "print completion-script-clink")
+		completionScriptFish  = flag.Bool("completion-script-fish", false, "print completion-script-fish")
 		targetsListPath       = flag.String("targets", "", "targets list file")
+		showVersion           = flag.Bool("version", false, "print version information")
 	)
 
 	flag.Parse()
+
+	if *showVersion {
+		version := VERSION
+		if version == "" {
+			version = "dev"
+		}
+		if BUILDDATE != "" {
+			fmt.Fprintf(c.outStream, "%s version %s (%s)\n", appName, version, BUILDDATE)
+		} else {
+			fmt.Fprintf(c.outStream, "%s version %s\n", appName, version)
+		}
+		return nil
+	}
 
 	if *completionScriptBash {
 		handleCompletionScriptBash(*targetsListPath)
@@ -41,6 +56,11 @@ func (c *cli) Run(args []string) error {
 
 	if *completionScriptClink {
 		handleCompletionScriptClink(*targetsListPath)
+		return nil
+	}
+
+	if *completionScriptFish {
+		handleCompletionScriptFish(*targetsListPath)
 		return nil
 	}
 
@@ -60,7 +80,7 @@ func (c *cli) Run(args []string) error {
 		flagCompleteMap["target"] = targets
 	}
 
-	fmt.Printf("%s\n", completionBash(flag.Args()))
+	fmt.Printf("%s\n", completeArgs(flag.Args()))
 
 	return nil
 }